Renamed ctapipe/calib/camera/calibrators.py to ctapipe/calib/camera/dl1.py
