+++ ctapipe/calib/camera/dl1.py
